Don't exclude all rules/files by default
I _believe_ this was initially included because we wanted to take an
opt-in approach, and at the time of hound-python being developed this
was only achievable at the linter config level. We now have the concept
of beta languages in Hound, which are disabled at the linter level by
default.

Removing this default config means those who opt-in to Python linting
will get the default config of the linter itself, rather than nothing.